requirements.txt
anthropic>=0.39.0
websockets>=12.0
cryptography>=42.0
aiosqlite>=0.19.0
pyyaml>=6.0
